src/requirements.txt
PyQt6>=6.4.0
opencv-python>=4.7.0
Pillow>=9.5.0
numpy>=1.24.0
pytest>=7.3.0
pytest-qt>=4.2.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
numba>=0.57.0
psutil>=5.9.0
memory-profiler>=0.60.0
line-profiler>=4.0.0